export available_encodings
