+++ server/Cargo.toml
@@ -6,12 +6,13 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-bytes = "1"
-futures-util = { version = "0.3", default-features = false }
-http-body-util = "0.1"
-hyper-util = { version = "0.1", features = ["full"] }
-hyper = { version = "1", features = ["full"] }
-serde_json = "1"
-serde = { version = "1.0", features = ["derive"] }
-tokio-util = "0.7.10"
-tokio = { version = "1", features = ["full"] }
+bytes = { workspace = true}
+config = { path = "../config" }
+futures-util = { workspace = true}
+http-body-util = { workspace = true}
+hyper-util = { workspace = true}
+hyper = { workspace = true}
+serde_json = { workspace = true}
+serde = { workspace = true}
+tokio-util = { workspace = true}
+tokio = { workspace = true}
